Change naming in a few places, add some comments, tidy up
